# crates/genomehubs-api/target/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GenomeHubs API",
    "description": "GenomeHubs Application Programming Interface.\n\nProvides a low-level API to retrieve records, look up taxa, search and generate reports across all taxon, assembly, and sample data.\n\nSee the [GenomeHubs documentation](https://genomehubs.org) for further details.",
    "contact": {
      "name": "GenomeHubs",
      "url": "https://genomehubs.org"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/genomehubs/genomehubs/blob/main/LICENSE"
    },
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v3/search/batch": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Fetch records for multiple search queries in a single request",
        "description": "Execute up to 100 search queries simultaneously using a single POST. Reduces network round-trips for parallel searches.",
        "operationId": "post_search_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchBatchRequest"
              },
              "examples": {
                "Two searches": {
                  "summary": "Search Mammalia and Insecta in parallel",
                  "value": {
                    "queries": [
                      {
                        "params_yaml": "size: 5\nfields: genome_size\ntaxonomy: ncbi\n",
                        "query_yaml": "index: taxon\nquery: tax_tree(Mammalia)\n"
                      },
                      {
                        "params_yaml": "size: 5\nfields: genome_size\ntaxonomy: ncbi\n",
                        "query_yaml": "index: taxon\nquery: tax_tree(Insecta)\n"
                      }
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchBatchResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiStatus": {
        "type": "object",
        "description": "Uniform status block present in every v3 API response.\n\nMetadata-only endpoints (taxonomies, ranks, indices) omit `hits` and `took`.\nQuery endpoints always populate all four fields.",
        "required": [
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "hits": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "took": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "SearchBatchItem": {
        "type": "object",
        "required": [
          "query_yaml",
          "params_yaml"
        ],
        "properties": {
          "params_yaml": {
            "type": "string"
          },
          "query_yaml": {
            "type": "string"
          }
        }
      },
      "SearchBatchRequest": {
        "type": "object",
        "required": [
          "searches"
        ],
        "properties": {
          "searches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchBatchItem"
            },
            "description": "Array of queries to search in batch (max 100)."
          }
        }
      },
      "SearchBatchResponse": {
        "type": "object",
        "required": [
          "status",
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchBatchResultItem"
            },
            "description": "Per-query search results."
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        }
      },
      "SearchBatchResultItem": {
        "type": "object",
        "required": [
          "status",
          "count",
          "hits"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of actual documents returned in hits array",
            "minimum": 0
          },
          "hits": {
            "type": "array",
            "items": {},
            "description": "Array of document results"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        }
      },
...
    }
  },
...
}